WP Plugin E2E tests: Create test site for e2e tests
Summary:
We need to do following things to enable automated e2e tests for wordpress plugin:
1. Create a test site for e2e testings
2. dockerize test site
3. Add build script to build and publish site image
4. Create E2E tests and create command start the tests
5. push images in FBCCastle

This the first step to create a test site. We are poring the initial wordpress site code to here first and make change on top of it later. This diff is just a part of wordpress source code as the code fils are too many so that we can't arc diff them in one diff.

Differential Revision: D13742661

fbshipit-source-id: a4c42293