Removing the warning message for plugin upgrade
Summary:
Removing the message - "***ATTENTION: Due to a known issue, upgrading this plugin may result in plugin deactivation (especially if you are upgrading from version 1.7.10 - 1.7.19). Please go to your site's 'Plugins' tab, refresh the page, and activate the plugin if it's deactivated to prevent losing your future pixel events.***"

It is not relevant anymore and it spoils the user experience.

Reviewed By: mynameisram

Differential Revision: D18977829

fbshipit-source-id: 614c37abbb6d05dfab87611a3f4a0e3ff081250a