Use http_referer as the event_source_url when dealing with forms
Summary:
Currently, the event source url sent doesn't correspond to the url in the browser for the form based plugins we support. This is because when the form is submitted, there is another call on the backend which processes it and url displayed on the browser becomes the HTTP_REFERER.

With this change, we'll take HTTP_REFERER or the REQUEST_URI whichever is expected to have the correct value.

Reviewed By: HeyMultiverse

Differential Revision: D21101320

fbshipit-source-id: 0faf76c9ddb4c62cf2c8bd4d8b58910f8e703e68