Support sending phone number back to server from CalderaForm for WordPress integration
Summary:
Support sending phone number back to server from CalderaForm for WordPress integration

If there are both better and basic version of the phone number attributes, take the better version. Fallback to the basic version if that's empty.

If there are multiple phone numbers, just take the first extracted one.

Reviewed By: vinothsa

Differential Revision: D21832467

fbshipit-source-id: 3db55a8e9e1fbe5d739b0cc6d47e49f7e8e1a676