Sending AddToCart CAPI event from EasyDigitalDownloads
Summary:
New methods are added in FacebookWordpressEasyDigitalDownloads to send addtocart pixel/CAPI events with the same event id. These methods are hooked to specific EED actions.
The purchase forms used to send product information when you click add to cart include a hidden field with the event id, when useS2S is enabled. These forms appear when you see a single product or a grid of products, if you use the [downloads] shortcut from EED.
The AJAX request with AddToCart information is intercepted and a new CAPI event is created, with the same event id as the pixel event.
The JS code used to send the pixel event is changed, to read event id from purchase form.

Reviewed By: vinothsa

Differential Revision: D22770091

fbshipit-source-id: 52b84dd411268f8a3ce43257255f4bf50746d23f