Adding phone and address information from Order and Customer session in CAPI events for WooCommerce
Summary:
Billing address information of the order is used to fill user data, in purchase events.
Billing address information of the customer is used to fill user data, in addtocart and initiatecheckout events.
Order information extraction is done by calling object methods, instead of reading an array.
Country field is stored in ISO format for both order and customer information.
Unit test are changed as well.

Reviewed By: vinothsa

Differential Revision: D22859617

fbshipit-source-id: 28174cabeaf77ae3364e34a369c3746969d5eefb