Setting agent string in all the pixel events sent by the plugin
Summary:
The function fbq('set', 'agent', <agent_string>, <pixel_id>) is called before sending any pixel event rendered by the WP plugin, in order to tag the events correctly. This modification is inside PixelRenderer class.
Because of this change, all the plugins's test are updated.
The method mockFacebookWordpressOptions(), in FacebookWordpressTestBase, substitutes mockUseS2S(), because FacebookWordpressOptions is used by the tests to get: usePii, useS2S, pixelId and agent; and it's easier to initialize the mocked object passing an array with these values.

Reviewed By: vinothsa

Differential Revision: D23230523

fbshipit-source-id: be4fd9fb2b1a2798aa66baee4dfa5e8780f831dd