Sending ViewContent and AddToCart from WooCommerce at the same time
Summary: When a user clicks the add to cart button in a product page in WooCommerce, ViewContent and AddToCart events are triggered, but only ViewContent is sent. This happens because wp_async_task just executes the run_action method once. The solution is to save both events and send them in the wp_footer action. Other event types are sent as they are created.

Reviewed By: marksliva

Differential Revision: D24323469

fbshipit-source-id: d436c3e06c7af61cecc396c29583acf2e41d8af6