Sending AddToCart pixel events from WooCommerce
Summary:
When a user clicks AddToCart in product pages or category pages, a CAPI and a Pixel event with the same event id are created.
Instead of printing the pixel event in the function that creates the CAPI event, the fbq call is enqueued with wc_enqueue_js() in order to be printed in the footer. This change is needed because the action associated with add to cart is executed before pixel initialization.

Reviewed By: marksliva

Differential Revision: D24366409

fbshipit-source-id: f6badade91ace5867871927d25ba5894563b332f