Sending pixel event when AJAX is used to add to cart in WooCommerce
Summary:
An user can add items to the cart in category pages or product pages. An admin can select between using ajax or not to add items to cart in WooCommerce plugin settings. When AJAX is used, wp_footer action is not executed and CAPI/Pixel events are not fired.
The solution is to send the AddToCart event inmediately if AJAX is used, and add a wordpress filter that will print the fbq call in the page.

Reviewed By: marksliva

Differential Revision: D24439474

fbshipit-source-id: 63235712858f54b3c4b419986b427613172f5f24