Using usePii key in AAM settings when FBE is not installed
Summary:
In the previous version, the use_pii key was used to send user data in Pixel and Conversions API. In the new one, the settings will be read from the AAM endpoint, once a user completes FBE flow.
For existing installations, with valid pixel ids, the usePii key will be transferred to the AAM object, if the flag was turned on, all the AAM fields will be sent in events. After finishing the FBE flow, AAM settings will be read and fetched normally.
Making this change will prevent a drop in the match rate of this Partner integration, because it's possible that existing users don't know about AAM in Events Manager. And also, we will be honoring previous advertisers preferences.

Reviewed By: marksliva

Differential Revision: D25632716

fbshipit-source-id: f05302c2d5b896114f36083b688a03216c828ecf