Adding pixel id and access token validation before sending Conversions API events
Summary:
Pixel id and access token are validated, by checking if they are empty, before being used to send Conversions API events.
The validation is added because for existing users, that pasted a pixel id, but not an access token, each time an event is created, Graph API exceptions will be thrown, because the AT is empty.

Reviewed By: davit-y

Differential Revision: D25799236

fbshipit-source-id: a9efb03e51647af6de8aaff9390a26f7c72be7b3