Using JSON functions instead of serialization to send events async
Summary:
Instead of using serialization functions in the class ServerEventAsyncTask, json functions are used to send events asynchronously. The method normalize from the Event class is used to convert an event to JSON string in prepare_data; and the event is rebuilt in the run_action method, using classes from the business sdk. Because the keys in the normalized data are different from the keys used in the constructors, they are parsed in run_action.

This was done because using php serialization functions  is unsafe to use with user requests (https://www.php.net/manual/en/function.unserialize.php) as demonstrated by the white hat researcher in the associated task.

Reviewed By: vinothsa

Differential Revision: D25787153

fbshipit-source-id: 95572cf59ab866ceea02c38a350215fe2f80c570