Fix CAPI Quality for Wordpress
Summary: CAPI quality for WordPress has dropped from ~85% to ~67% as more advertisers install the latest version of the extension with the open-bridge feature. This root cause has been identified as a technical issue in how the WordPress plugin decodes the pixel events payload into CAPI. Some PII is dropped. This diff is meant to fix that.

Differential Revision: D40760101

fbshipit-source-id: a8acc9bb00757b8ea201de285fbe5a01d00ae9cc