Remove async processing from openBridge events since they are already done via Ajax
Summary: Openbridge events are processed through AJAX, and the browser does not wait for a response. Consequently, any delay in the connection to send events to the Facebook server will not impact the speed of the website. So we are making this change to stop using AsyncTask for OpenBridge events

Reviewed By: vahidkay-meta

Differential Revision: D46501597

fbshipit-source-id: 04d374b20192d8b7d5cf9314a2c825f7e8643bad