Remove plugin settings matching old DB key on uninstall (#67)
Summary:
This changeset makes sure that both database entries containing the plugin settings are removed when uninstalling the plugin. Previously, only the current setting DB entry was being removed when uninstalling the plugin.

Because the plugin copies values from the old DB entry (`facebook_config`) to the new one (`facebook_business_extension_config`) when being run, the plugin never returns to an "empty" state when reinstalled on a machine that had an older version of the plugin which used the `facebook_config` options DB key.

Pull Request resolved: https://github.com/facebookincubator/Facebook-Pixel-for-Wordpress/pull/67

Reviewed By: tzahgr

Differential Revision: D68435405

Pulled By: vahidkay-meta

fbshipit-source-id: 001dcb230b49a3d3cbd77dae423a78fb4fac87e0