Read fbc & fbp from databag instead of relying on cookies (#79)
Summary:
When a cookie is set on the client side, the backend does not see that cookie on the first load. Therefore, fbc & fbp are always missed on the first load.

In this PR, we are reading those values from the payload that's sent to the backend by OpenBridge. This does not need to be done through setting cookies and instead we can rely on the payload itself.

This can be tested by:
- Putting debug logs in either class-servereventfactory.php to see if fbp & fbc values exist in the user_data
- Checking the match keys that are sent to EventsManager on the first call ( must make sure not to send an event twice, or load the page more than once )

Pull Request resolved: https://github.com/facebookincubator/Facebook-Pixel-for-Wordpress/pull/79

Reviewed By: akpurifb

Differential Revision: D69179107

Pulled By: vahidkay-meta

fbshipit-source-id: bb50878fcf9d72853cfd73f08765712e1fe6c6df