Set up CAPI in the same script block as init
Summary:
The Meta Pixel fires CAPI events using an openbridge connection. This connection is set up with a call to configure it directly. If this call happens after to the pixel being initialized, then the pixel will not re-configure itself to set up the openbridge plugin properly.

This diff updates the config to be set in the same script block as the init, eliminating the race condition that allows this case to happen.

Differential Revision: D73067691