Set up CAPI in the same script block as init (#96)
Summary:
The Meta Pixel fires CAPI events using an openbridge connection. This connection is set up with a call to configure it directly. If this call happens after to the pixel being initialized, then the pixel will not re-configure itself to set up the openbridge plugin properly.

This diff updates the config to be set in the same script block as the init, eliminating the race condition that allows this case to happen.

Pull Request resolved: https://github.com/facebookincubator/Facebook-Pixel-for-Wordpress/pull/96

Test Plan:
Tests updated
|Before|After|
|--|--|
| {F1977152428} | {F1977152398}|

Reviewed By: marksliva

Differential Revision: D73068479

Pulled By: jarretth

fbshipit-source-id: 0162641847d048916c22040353865d71d7c850c6